Reduce duplicated code
