+++ annot8-common/annot8-common-processing/src/main/java/io/annot8/common/processing/filters/GroupFilters.java
@@ -8,7 +8,6 @@ import io.annot8.api.filters.Filter;
 import io.annot8.api.filters.NotFilter;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -19,19 +18,19 @@ public class GroupFilters {
   }
 
   public static Filter<Group> byType(String type) {
-    return new TypeFilter(type);
+    return new TypeFilters.TypeFilter<>(type);
   }
 
   public static Filter<Group> byProperty(String key) {
-    return new HasPropertyFilter(key, null, null);
+    return new PropertyFilters.HasPropertyFilter<>(key, null, null);
   }
 
   public static Filter<Group> byProperty(String key, Class<?> clazz) {
-    return new HasPropertyFilter(key, clazz, null);
+    return new PropertyFilters.HasPropertyFilter<>(key, clazz, null);
   }
 
   public static Filter<Group> byProperty(String key, Object value) {
-    return new HasPropertyFilter(key, null, value);
+    return new PropertyFilters.HasPropertyFilter<>(key, null, value);
   }
 
   public static Filter<Group> hasRoles(String... roles) {
@@ -48,47 +47,6 @@ public class GroupFilters {
 
   public static Filter<Group> and(Filter<Group>... filters) {
     return new AndFilter<>(filters);
-  }
-
-  public static class TypeFilter implements Filter<Group> {
-    private final String type;
-
-    public TypeFilter(String type) {
-      this.type = type;
-    }
-
-    @Override
-    public boolean test(Group group) {
-      return group.getType().equals(type);
-    }
-  }
-
-  // TODO: Same as ANnotation (WithProperties?)
-  public static class HasPropertyFilter implements Filter<Group> {
-
-    private final String key;
-    private final Class<?> valueClass;
-    private final Object value;
-
-    public HasPropertyFilter(String key, Class<?> valueClass, Object value) {
-      this.key = key;
-      this.valueClass = valueClass;
-      this.value = value;
-    }
-
-    @Override
-    public boolean test(Group group) {
-      Optional<Object> o = group.getProperties().get(key);
-      if (o.isEmpty()) {
-        return false;
-      } else if (value != null) {
-        return o.get().equals(value);
-      } else if (valueClass != null) {
-        return valueClass.isInstance(o.get());
-      } else {
-        return true;
-      }
-    }
   }
 
   public static class HasRolesFilter implements Filter<Group> {
